pyproject.toml
[tool.poetry]
name = "hierarchical-data"
version = "0.1.0"
description = "app that shows the heirarchical structure of folders and files"
authors = ["Michael DeMory <mjdemory2891@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
django-mptt = "^0.11.0"
django = "^3.1.1"
ipdb = "^0.13.3"

[tool.poetry.dev-dependencies]
pylint = "^2.6.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
